+++ src/app/views/common/registry/popups.tsx
@@ -4,7 +4,7 @@ export const popups = new Map<string, any>([
   ['share-query', lazy(() => import('../../query-runner/query-input/share-query/ShareQuery'))],
   ['theme-chooser', lazy(() => import('../../main-header/settings/ThemeChooser'))],
   ['preview-collection', lazy(() => import('../../sidebar/resource-explorer/collection/PreviewCollection'))],
-  ['full-permissions', lazy(() => import('../../query-runner/request/permissions/PanelList'))]
+  ['full-permissions', lazy(() => import('../../query-runner/request/permissions/Permissions.Full'))]
 ]);
 
 export type PopupItem =
