+++ src/app/views/query-runner/request/permissions/index.ts
@@ -1,5 +1,5 @@
-import { Permission } from './Permission';
+import { Permissions } from './Permissions.Query';
 
 export {
-  Permission
+  Permissions
 };
